:sparkles: refactor: move update learner progress to separate file
