+++ libs/functions/bot-engine/main/src/lib/services/process-message/process-message.service.ts
@@ -13,9 +13,6 @@ import { NextBlockFactory } from '../next-block/next-block.factory';
 import { CursorDataService } from '../data-services/cursor.service';
 import { ConnectionsDataService } from '../data-services/connections.service';
 import { BlockDataService } from '../data-services/blocks.service';
-import { EnrolledUserDataService } from '../data-services/enrolled-user.service';
-import { BotModuleDataService } from '../data-services/botmodules.service';
-import { StoriesDataService } from '../data-services/stories.service';
 
 import { BotMediaProcessService } from '../media/process-media-service';
 import { ProcessInputFactory } from '../process-input/process-input.factory';
@@ -23,6 +20,8 @@ import { OperationBlockFactory } from '../process-operation-block/process-operat
 import { assessUserAnswer } from '../process-operation-block/block-type/assess-user-answer';
 import { SurveyService } from '../process-operation-block/block-type/survey-service';
 import { FallBackBlockService } from '../next-block/block-type/fallback-block.service';
+
+import { updateLearnerProgress } from '../../utils/updateLearnerProgress.util';
 
 
 export class ProcessMessageService
@@ -95,7 +94,7 @@ export class ProcessMessageService
     this.sideOperations.push(inputPromise);
 
     // upodate leaner progrress
-    const updateLearnersProgressPromise = this._updateLearnerProgress(currentStory, lastBlock, endUser, tools, orgId);
+    const updateLearnersProgressPromise = updateLearnerProgress(currentStory, lastBlock, endUser, tools, orgId);
 
     this.sideOperations.push(updateLearnersProgressPromise);
 
@@ -204,57 +203,6 @@ export class ProcessMessageService
 
     return updatedPosition;
   }
-
-  /** update the enrolled user's progress with the just completed story(lesson's) details*/
-  private async _updateLearnerProgress (currentStory: string, lastBlock:StoryBlock, endUser:EndUser, tools:HandlerTools, orgId:string) {
-    const enrolledDataServ = new EnrolledUserDataService(tools, orgId);
-    const botModDataService = new BotModuleDataService(tools, orgId);
-    const storiesDataService = new StoriesDataService(tools, orgId);
-
-    const enrolledUser = await enrolledDataServ.getEnrolledUser(endUser.enrolledUserId ?? '');
-
-    const parentModule = (await storiesDataService.getStory(currentStory))?.parentModule;
-    const parentCourse = (await botModDataService.getBotModule(parentModule))?.parentBot;
-
-    if (!enrolledUser) return;
-
-    if (enrolledUser.courses == null) enrolledUser.courses = [];
-
-    // Find or create the course
-    let theCourse = enrolledUser.courses.find(course => course.courseId === parentCourse);
-
-    if (!theCourse) {
-      theCourse = { courseId: parentCourse, modules: [] };
-      enrolledUser.courses.push(theCourse);
-    }
-
-    // Find or create the module
-    let theModule = theCourse.modules.find(mod => mod.moduleId === parentModule);
-
-    if (!theModule) {
-      theModule = { moduleId: parentModule, lessons: [] };
-      theCourse.modules.push(theModule);
-    }
-
-    let theLesson = theModule.lessons.find(lesson => lesson.lessonId === currentStory);
-
-    if (!theLesson) {
-      theLesson = { lessonId: currentStory, blocks: [] };
-      theModule.lessons.push(theLesson);
-    }
-
-    const block = theLesson.blocks.find((blockId) => blockId === lastBlock.id)
-    if (!block) theLesson.blocks.push(lastBlock.id);
-
-    tools.Logger.log(() => `Updating enrolled user progress for ${endUser.enrolledUserId}`);
-    tools.Logger.log(() => `Current Story: ${currentStory}`);
-    tools.Logger.log(() => `Parent Course: ${parentCourse}`);
-    tools.Logger.log(() => `Parent Module: ${parentModule}`);
-    tools.Logger.log(() => `Lesson: ${JSON.stringify(theLesson)}`);
-
-    // Save changes
-    await enrolledDataServ.updateEnrolledUser(enrolledUser);
-  };
 
   public getSideOperations()
   {
